% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_clade}
\alias{filter_clade}
\title{Filter a vector of taxa to those belonging to a given clade}
\usage{
filter_clade(taxa, clade, verbose = FALSE)
}
\arguments{
\item{taxa}{A character vector of taxon names.}

\item{clade}{A character string giving the clade to filter by.}

\item{verbose}{Logical. If \code{TRUE}, prints progress messages. Default \code{FALSE}.}
}
\value{
A character vector of taxa that are members of the specified clade.
}
\description{
Given a vector of taxon names and a clade name, returns only those taxa
whose lineage includes the specified clade.
}
\examples{
\dontrun{
taxa <- c("Tyrannosaurus", "Triceratops", "Velociraptor",
          "Brachiosaurus", "Homo")
filter_clade(taxa, "Theropoda")
filter_clade(taxa, "Dinosauria")
}
}
