% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{print.taxodist_bundle}
\alias{print.taxodist_bundle}
\title{Print a taxodist bundle}
\usage{
\method{print}{taxodist_bundle}(x, ...)
}
\arguments{
\item{x}{A \code{taxodist_bundle} object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a taxodist bundle
}
