% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{read_taxodist_bundle}
\alias{read_taxodist_bundle}
\title{Read a portable taxodist JSON bundle}
\usage{
read_taxodist_bundle(file)
}
\arguments{
\item{file}{Path to a JSON file created by \code{\link[=write_taxodist_bundle]{write_taxodist_bundle()}}.}
}
\value{
A reconstructed \code{taxodist_bundle} object.
}
\description{
Read a portable taxodist JSON bundle
}
